requirements.txt
langchain-community
faiss-cpu
fastapi
requests